+++ tests/helpers/dbt_tests/cases/jaffle_customers_incremental_new_column.sql
