+++ tests/helpers/dbt_tests/cases/jaffle_customers_with_error.sql
