+++ tests/helpers/dbt_tests/utils.py
@@ -7,7 +7,7 @@ from dlt.common.git import clone_repo
 from dlt.helpers.dbt.exceptions import DBTNodeResult
 
 JAFFLE_SHOP_REPO = "https://github.com/dbt-labs/jaffle_shop.git"
-TEST_CASES_PATH = "./tests/dbt_runner/cases/"
+TEST_CASES_PATH = "./tests/helpers/dbt_tests/cases/"
 
 
 def load_test_case(case: str) -> str:
